N501Y変異株スクリーニングの実施状況: Rename components/VariantChart.vue to components/index/CardsReference/Variant/Chart.vue
