+++ components/index/CardsReference/Variant/Chart.vue
